+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Added
 *   Parameters `paymentReference` and `invoiceId` to `Authorization` and `Payout`.
 *   The SDK now supports the webhook event `payout`.
+*   Example for Flexipay direct.
 
 ### Changed
 *   The webhook tests now cover all supported events.