+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 *   Parameters `paymentReference` and `invoiceId` to `Authorization` and `Payout`.
 *   The SDK now supports the webhook event `payout`.
 *   Example for Flexipay direct.
+*   Email parameter to `Paypal` payment type.
 
 ### Changed
 *   The webhook tests now cover all supported events.