+++ CHANGELOG.md
@@ -4,7 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [1.2.X.X][1.2.X.X]
+
 ### Added
+*   Example for Flexipay direct.
 *   The SDK now supports the webhook event `payout`.
 
 ### Changed